+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Allow a `requests.Session` object to be passed to `WsiDicomWebClient.create_client()`.
+
+### Changed
+
+- Refactored `WsiDicomWebClient.__init__()` to take a `DICOMwebClient` instance, and moved the old `__init__()` method into `WsiDicomWebClient.create_client()`.
+- Make some arguments optional for `WsiDicomWebClient.create_client()`.
+- Removed `WsiDicomFileClient` since it is no longer needed.
+
 ## [0.12.0] - 2023-10-04
 
 ### Changed